Move units and items into world
