+++ android/jni/src/Android.mk
@@ -9,11 +9,9 @@ LOCAL_SRC_FILES := ../../../../libraries/SDL/src/main/android/SDL_android_main.c
 FILE_LIST := $(wildcard $(LOCAL_PATH)/core_components/*.cpp)
 FILE_LIST += $(wildcard $(LOCAL_PATH)/physics/*.cpp)
 FILE_LIST += $(wildcard $(LOCAL_PATH)/commands/*.cpp)
-FILE_LIST += $(wildcard $(LOCAL_PATH)/units/*.cpp)
-FILE_LIST += $(wildcard $(LOCAL_PATH)/units/*/*.cpp)
-FILE_LIST += $(wildcard $(LOCAL_PATH)/items/*.cpp)
-FILE_LIST += $(wildcard $(LOCAL_PATH)/items/*/*.cpp)
 FILE_LIST += $(wildcard $(LOCAL_PATH)/world/*.cpp)
+FILE_LIST += $(wildcard $(LOCAL_PATH)/world/*/*.cpp)
+FILE_LIST += $(wildcard $(LOCAL_PATH)/world/*/*/*.cpp)
 FILE_LIST += $(wildcard $(LOCAL_PATH)/ui_components/*.cpp)
 
 LOCAL_SRC_FILES += $(FILE_LIST:$(LOCAL_PATH)/%=%)
