+++ android/jni/src/commands/MoveCommand.h
@@ -10,7 +10,7 @@
 #include "oxygine-framework.h"
 
 #include "commands/DirectionalCommand.h"
-#include "units/Unit.h"
+#include "world/units/Unit.h"
 #include "world/World.h"
 
 using namespace oxygine;
