--- android/jni/src/items/items.h
@@ -1,18 +1,0 @@
-/**
- * items.h
- *
- * Description: List of items for ease.
- */
-
-#ifndef SWORD_FIGHTER_ITEMS_H
-#define SWORD_FIGHTER_ITEMS_H
-
-#include "items/armor/EmptyItem.h"
-#include "items/armor/EmptyHead.h"
-#include "items/armor/EmptyTorso.h"
-#include "items/armor/EmptyLegs.h"
-
-#include "items/weapons/RawSword.h"
-#include "items/weapons/Fist.h"
-
-#endif //SWORD_FIGHTER_ITEMS_H
