+++ android/jni/src/world/items/Item.h
