+++ android/jni/src/world/items/armor/EmptyLegs.cpp
@@ -2,11 +2,11 @@
 // Created by cvalenza on 2/17/17.
 //
 
-#include "EmptyHead.h"
+#include "world/items/armor/EmptyLegs.h"
 
-EmptyHead::EmptyHead()
+EmptyLegs::EmptyLegs()
 {
-	item_type = Wearable::ItemType::HEADGEAR;
+	item_type = Wearable::ItemType::PANTS;
 	effect_type = Wearable::ItemEffect::NONE;
 	init("empty", Item::WEARABLE);
 }
