+++ android/jni/src/world/items/items.h
@@ -1,0 +1,18 @@
+/**
+ * items.h
+ *
+ * Description: List of items for ease.
+ */
+
+#ifndef SWORD_FIGHTER_ITEMS_H
+#define SWORD_FIGHTER_ITEMS_H
+
+#include "world/items/armor/EmptyItem.h"
+#include "world/items/armor/EmptyHead.h"
+#include "world/items/armor/EmptyTorso.h"
+#include "world/items/armor/EmptyLegs.h"
+
+#include "world/items/weapons/RawSword.h"
+#include "world/items/weapons/Fist.h"
+
+#endif //SWORD_FIGHTER_ITEMS_H
